Update to docker files (#510)
* Added chdir to run.sh

When run as a docker image, it fails saying it is unable to find config file.  This is because "run.sh" executes in "/" (and so does "codecept.js"), but config file is in "/tests/".  This commit changes directory before executing codecept.js.

* Updated run.sh for full path to codecept.js

* Removed directory name from codecept.js in run.sh

* Replaced GREP env variable with CODECEPT_ARGS

Replaced GREP env variable with CODECEPT_ARGS to allow more flexibility in run.sh.