Fix attachments usage for `mocha-junit-reporter`
The current implementation checks for the `attachments` option in the
wrong location, which requires the user to set `attachments:true` in two
places, the current location and within the `options` object.

This changes the check and docs to use the `options` object which is
what `mocha-junit-report` is actually looking for.